requirements.txt
Flask==3.0.0
pytest==8.4.2
PyYAML==6.0.3
psycopg2-binary==2.9.10
psycopg==3.2.3